[BE] Add Authentication and Authorization for Backend

+++ MyDiary.Application/Diary/Commands/DeleteDiary/DeleteDiaryCommandHandler.cs
@@ -1,5 +1,7 @@
 ﻿using MediatR;
+using Microsoft.Extensions.Logging;
 using MyDiary.Application.Exceptions;
+using MyDiary.Domain.Interfaces;
 using MyDiary.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -9,7 +11,10 @@ using System.Threading.Tasks;
 
 namespace MyDiary.Application.Diary.Commands.DeleteDiary
 {
-    public class DeleteDiaryCommandHandler(IDiaryRepository diaryRepository, IUnitOfWork unitOfWork) : IRequestHandler<DeleteDiaryCommand>
+    public class DeleteDiaryCommandHandler(IDiaryRepository diaryRepository, 
+        IUnitOfWork unitOfWork,
+        ILogger<DeleteDiaryCommandHandler> logger,
+        IDiaryAuthorizationService diaryAuthorizationService) : IRequestHandler<DeleteDiaryCommand>
     {
         public async Task Handle(DeleteDiaryCommand request, CancellationToken cancellationToken)
         {
@@ -21,6 +26,11 @@ namespace MyDiary.Application.Diary.Commands.DeleteDiary
                 var diary = await diaryRepository.GetById(request.DiaryId);
                 if (diary == null)
                     throw new NotFoundException(nameof(diary), request.DiaryId.ToString());
+
+                logger.LogWarning("Remove diary: {diaryId}", request.DiaryId);
+
+                if (!diaryAuthorizationService.Authorize(diary, Domain.Constants.ResourceOperation.Delete))
+                    throw new ForbidException();
 
                 await unitOfWork.BeginTransactionAsync();
                 await diaryRepository.Delete(diary);
